Add files via upload
Text files for players to have to go along with the project.  Includes a list of updated Djinn/Djinn charms, along with where to find them; a list of items Monsters can drop when defeated; and a basic FAQ of how to progress through the game, as the dialogue that would often help the player navigate through it has been removed.