Add files via upload
Worked on this because of I noticed some quick needed fixes on stream.  In hindsight, the classes for jack adept were just fine, I was just streaming the wrong patch.  Though I decided to up the level-up requirement for gravity while I was in the editor and make a minor change to a boss fight, I've been meaning to do for a while but didn't because I could have sworn last patch was going to be the last patch of version 1. Feel pretty confident THIS will be the last patch of version 1 unless something SUPER game breaking comes up.